Move lufa descriptor to protocol/usb_descriptor
